backend/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
google-cloud-secret-manager